requirements.txt
noise
numpy
pyaudio
soundfile
webrtcvad
speechbrain
sentence-transformers
networkx
openai
spacy
requests
PyQt5
toml
modelscope
live2d-py
sounddevice
pyautogui
torchaudio
hashlib